Refactor: Separate concerns and improve code structure
This commit addresses several refactoring tasks to improve the maintainability, readability, and reusability of the codebase.

- Centralized constants: Moved navigation and footer links from components into a dedicated `constants/navigation.tsx` file.
- Separated API services: Extracted the contact form submission logic into a `services/contactService.ts` file.
- Created custom hooks: Abstracted the mobile menu state and logic from the Header component into a `hooks/useMobileMenu.ts` hook.
- Split components: Decomposed the large post client component into smaller, more focused `PostHeader` and `PostNavigation` components.
