+++ src/hooks/useMobileMenu.ts
@@ -1,0 +1,19 @@
+import { useState } from "react";
+
+export const useMobileMenu = () => {
+  const [isMenuOpen, setIsMenuOpen] = useState(false);
+
+  const toggleMenu = () => {
+    setIsMenuOpen((prev) => !prev);
+  };
+
+  const closeMenu = () => {
+    setIsMenuOpen(false);
+  };
+
+  return {
+    isMenuOpen,
+    toggleMenu,
+    closeMenu,
+  };
+};
